Ask in one place whether a tuple stands at a point
The walk that answers it lived inside the measurement, so only what is
already written could be asked. A generation that wanted the same
question about a row it had just composed would have had to write the
rule a second time, and the two would agree until one of them moved.

`StandingAtAPoint` is that walk, put where both can reach it. Nothing
about how a point is met changed: the readings of a row are enumerated
the same way, the quantity is still the one thing that says what a value
at the point is, and a line a fork drew is still met by getting the
comparison to answer.

Its answer now tells a row nobody watched from a row that ran and did
not reach the comparison. Both used to arrive as an empty account and
came out as a miss. The measurement still counts the first as a miss,
which is what it was — and says so where it does it, because what the
rows went without is what it weakens its verdict by. A reader for which
the two differ can now ask.

Two members of the walk were dead: an accessor nothing called, and an
arm of the answer nothing produced.
